# Hash 1049