## Overview

`
$  benchcmp xxhash sha1
benchmark                      old ns/op      new ns/op      delta
BenchmarkFragment_Blocks-8     1527160039     2351544072     +53.98%
`
Fixes #1049

NOTE: after upgrading Pilosa with this change, do not perform a rolling restart. The whole cluster needs to come down, get the new binary, then come back up.

## Pull request checklist

- [x] I have read the [contributing guide](https://github.com/pilosa/pilosa/blob/master/CONTRIBUTING.md).
- [x] I have agreed to the [Contributor License Agreement](https://cla-assistant.io/pilosa/pilosa).
- [x] I have updated the [documentation](https://github.com/pilosa/pilosa/tree/master/docs).
- [x] I have resolved any merge conflicts.
- [x] I have included tests that cover my changes.
- [x] All new and existing tests pass.

## Code review checklist
This is the checklist that the reviewer will follow while reviewing your pull request. You do not need to do anything with this checklist, but be aware of what the reviewer will be looking for.

- [x] Ensure that any changes to external docs have been included in this pull request.
- [x] If the changes require that minor/major versions need to be updated, tag the PR appropriately.
- [x] Ensure the new code is [properly commented](https://github.com/golang/go/wiki/CodeReviewComments#doc-comments) and follows [Idiomatic Go](https://dmitri.shuralyov.com/idiomatic-go).
- [x] Check that tests have been written and that they cover the new functionality.
- [x] Run tests and ensure they pass.
- [x] Build and run the code, performing any applicable integration testing.